requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.2.1
pre-commit==4.2.0
pylint==3.3.6
pytest==7.3.1
pytest-html==4.1.1
PyYAML==6.0.1
selenium==4.15.1
